Lint (#49)
* include all folders for lint

* Add automatic lint fixes

These files were changes as a result of `npm run lint:fix`.

* Add imported classes to globals

Resolve issues of type no-undef from eslint.

* Add ignorewarning for warnings about localize

Ignore warnings caused due to usage of the localize function.

* Fix lang terms in locales.json

* remove unused input argument in _showMissedEntity

This was causing a lint error and was fixed by removing the unused input
argument.

* Resolve lint errors caused due to indentation

The script blocks in these topic-name and module-name had incorrect
indentation.

* Resolver other lint errors

Resolve lint errors about use of `this` keyword

* Remove `global` comments

The variables declared global in these comments were moved to
.eslintrc.json

* Add global variables to .eslintrc.json

This will eliminate the need to write global comments everytime we want
to use a class that's declared outside of the file

* Add npm run test to scripts in .travis.yml

* Fix lint issue in iterator.html

* Fix tests by changing element names

Some of the tests used `button` instead of `d2l-button` and `mixed-link`
isntead of `link`. This commit fixes those issues.

* Remove all /* global comments

We no longer need these comments, the required classes and functions can
just be added to .eslintrc.json.

* Add entity as argument to render

entity was removed due to a lint issue. I've added a comment to ignore
the lint rule that was causing the issue and have brought back entity as
an argument to _render().